E-commerce Data Table Structure

The functions in this package operate on a data.table containing e-commerce transaction data.

Format

A data .table with roughly 300.000 observations of unique invoice items The table must have the following columns:

invoice_no (integer): Invoice number

stock_code (character): Product stock code

description (character): Product description

quantity (integer): Quantity sold

unit_price (numeric): Price per unit

customer_id (integer): Customer ID

country (character): Country of purchase

sales (numeric): Total sales amount

date (numeric): Date of transaction

year (integer): Year of transaction

month (integer): Month of transaction

weekday (integer): Weekday of transaction

hour (integer): Hour of transaction

Source

https://archive.ics.uci.edu/ml/datasets/online+retail