E-commerce Data Table Structure

The functions in this package operate on a data.table containing e-commerce transaction data. The table must have the following columns:

Usage

calculate_market_share_by_country(x)

Arguments

x

A data.table following the structure defined in the "E-commerce Data Table Structure" section.

Value

A data.table with the market share percentage per country.

Details

invoice_no (integer): Invoice number

stock_code (character): Product stock code

description (character): Product description

quantity (integer): Quantity sold

unit_price (numeric): Price per unit

customer_id (integer): Customer ID

country (character): Country of purchase

sales (numeric): Total sales amount

date (numeric): Date of transaction

year (integer): Year of transaction

month (integer): Month of transaction

weekday (integer): Weekday of transaction

hour (integer): Hour of transaction Calculate Market Share by Country

This function calculates the percentage of total sales per country. Countries with less than 0.9% of total sales are grouped under "other".